+++ commons/persistence.js
@@ -3,6 +3,7 @@ const broker = require('../commons/broker.js');
 const tenancy = require('../commons/tenancy.js');
 const factory = require('../commons/factory.js');
 const bcrypt = require('bcryptjs');
+const { passwordStrength } = require('check-password-strength');
 
 exports.get = async function (id, field, model, select) {
     try {
@@ -50,7 +51,7 @@ exports.getList = async function (filter, sort, select, page, limit, restriction
 }
 
 const postOne = async function (body, model, tenant) {
-    if (body.password) if (tenant.passwordhash == true) body.password = bcrypt.hashSync(body.password, 8);
+    if (body.password) body.password = checkPassword(body.password,tenant.passwordhash);
     const resource = await (new model(body)).save();
     if (model.modelName == 'Measurement') {
         broker.publish('device-' + body.device, body.device, body);
@@ -67,7 +68,7 @@ const postList = async function (body, model, tenant) {
     for (let [i, element] of body.entries()) {
         try {
             element.owner = body.owner;
-            if (element.password) if (tenant.passwordhash == true) element.password = bcrypt.hashSync(element.password, 8);
+            if (element.password) element.password = checkPassword(element.password,tenant.passwordhash);
             const resource = await (new model(element)).save()
             if (model.modelName == "Measurement") {
                 broker.publish('device-' + resource.device, resource);
@@ -103,7 +104,7 @@ exports.update = async function (body, fields, resource, model, tenant, query, r
 
     for (let field of fields) {
         if (typeof body[field] != 'object' && body[field]) {
-            if (field == 'password') if (tenant.passwordhash == true) body[field] = bcrypt.hashSync(body[field], 8);
+            if (field == 'password') body[field] = checkPassword(body[field],tenant.passwordhash);
             resource[field] = body[field]; continue;
         }
         if (typeof body[field] == 'object' && body[field]) {
@@ -262,4 +263,11 @@ const modifyEmbeddedResourceList = async function (list, resource, field, identi
     }
     resource[field] = [...new Set(resource[field])];
     return [true, report];
+}
+
+const checkPassword = function (password,passwordhash) {
+    const details = passwordStrength(password);
+    if (details.id < process.env.MIN_PASSWORD_STRENGTH) throw new Error('The password strength is ' + details.value + ', please choose a stronger password');//MIN_PASSWORD_STRENGTH:0=TOO WEAK; 1=WEAK; 2=MEDIUM; 3=STRONG
+    if (passwordhash === false || passwordhash === 'false') return password;
+    return bcrypt.hashSync(password, 8);
 }