+++ test/login-test.js
@@ -45,6 +45,26 @@ describe('/POST login', () => {
         res.body.message.should.contain(errors.authentication_error.message);
         res.body.details.should.contain('Incorrect username or password');
     });
+
+    it('it should not POST a login with expired password validity time', async () => {
+        await factory.createUser("test-username-1", "test-password-1",null,null,null,null,1,1500000000000);
+        const request = {username: "test-username-1", password: "test-password-1"};
+        const res = await chai.request(server).keepOpen().post('/v1/login').send(request);
+        res.should.have.status(errors.authentication_error.status);
+        res.body.should.be.a('object');
+        res.body.message.should.be.a('string');
+        res.body.message.should.contain(errors.authentication_error.message);
+        res.body.details.should.contain("Password validity expired, please reset password");
+    });
+
+    it('it should POST a login with validityPasswordDays equal to 0 like no expiration date', async () => {
+        await factory.createUser("test-username-1", "test-password-1",null,null,null,null,0,1500000000000);
+        const request = {username: "test-username-1", password: "test-password-1"};
+        const res = await chai.request(server).keepOpen().post('/v1/login').send(request);
+        res.should.have.status(200);
+        res.body.should.be.a('object');
+        res.body.should.have.property('token');
+    });
 });
 
 // Test the /PUT route
@@ -80,5 +100,16 @@ describe('/PUT login', () => {
         res.body.message.should.contain(errors.authentication_error.message);
         res.body.details.should.contain('maxAge exceeded');
         process.env.JWT_RENEW_EXPIRATIONTIME = temp;
+    });
+
+    it('it NOT should PUT a login with a correct token but expired password validity time', async () => {
+        const user = await factory.createUser("test-username-1", "test-password-1",null,null,null,null,1,1500000000000);
+        const request = {};
+        const res = await chai.request(server).keepOpen().put('/v1/login').set("Authorization", await factory.getUserToken(user)).send(request);
+        res.should.have.status(errors.authentication_error.status);
+        res.body.should.be.a('object');
+        res.body.message.should.be.a('string');
+        res.body.message.should.contain(errors.authentication_error.message);
+        res.body.details.should.contain("Password validity expired, please reset password");
     });
 });